requirements.txt
azure-ai-ml
azure-identity
pandas
python-dotenv
xlrd